requirements.txt
rhasspy-asr-pocketsphinx==0.1.0
rhasspy-silence==0.1.0
rhasspy-hermes==0.1.3
paho-mqtt==1.5.0
